requirements.txt
numpy==1.26.2
Pillow~=10.1.0
argparse~=1.4.0
opencv-python~=4.8.1.78
colormath~=3.0.0
colour-science~=0.4.3
